+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Adds a toast notification to provide feedback when copying a word with "Tap to Copy" turned on (ABY-14).
+
 ## [1.1] - 19-02-2022
 
 - Apple Watch Companion App: Generate sniglets on-the-fly and save them with the new app for Apple Watch.